VectorSearchMiddleware/requirements.txt
Flask
gunicorn
face_recognition
numpy
datasets
oracledb
Pillow
python-dotenv